Tidy up imports, whitespace and workspace.
General housekeeping
