+++ .gitignore
@@ -139,3 +139,4 @@ cloud-auth.json
 
 
 cruft
+staticfiles
