+++ pytest.ini
@@ -4,3 +4,5 @@ addopts = --reuse-db --maxfail=0
 python_files = tests.py test_*.py *_tests.py
 markers =
     only: "Convenience method, so we can run a focussed test in pytest-watch"
+    smoke_test: "Smoke test - for exercising external APIs"
+    object_storage: "Uses object storage"
